Fix bugs : -Profile -> Edit address -> Terlihat ketika mode edit dropdown dari kabupaten mengambil semua kabupaten yang ada di Indonesia bukan lagi berdasarkan provinsinya yaitu Jawa Barat.
-Profile -> Edit -> No telp dimasukan string -> hasilnya pop logout, sebaiknya ada handle validasi di formnya seperti yang ada di form price

-Profile -> New Address -> Belum ada validasi untuk alamat yang tidak diisi[integrity exception null column address]

-Add Delete Confirmation:
1.my profile
2.checkout
3.checkout guest

+++ resources/views/shop/checkout.blade.php
@@ -153,7 +153,7 @@
                                             <div class="cust-addr-phone">
                                                 {{ $lua->_receiver_phone }}
                                             </div>
-                                            <a href="shop/address/delete/{{ $lua->id }}">Delete</a>
+                                            <a href="shop/address/delete/{{ $lua->id }}" onclick="return confirm('Are you sure?')">Delete</a>
                                         </div>
                                     </div>
                                 @endforeach
